(fix) Caching api token (feat) reworked the scratchpad to better suit requirements
